Remove account hierarchy modeling
Ultimately accounts will need to be constructed as a hierarchy, but it's unclear at this point exactly what that will look like. To simplify things for now, removing the formal hierarchy and instead just using the name to define associations between accounts. Adding an IUpsertAccountService and updating account related tests to not rely on pre-seeded database values but instead add accounts as part of setting up the test.